+++ tomcat/src/main/java/org/apache/catalina/connector/Connector.java
@@ -4,7 +4,7 @@ import java.io.IOException;
 import java.io.UncheckedIOException;
 import java.net.ServerSocket;
 import java.net.Socket;
-import org.apache.coyote.RequestHandler;
+import org.apache.catalina.RequestHandler;
 import org.apache.coyote.http11.Http11Processor;
 import org.apache.coyote.http11.HttpRequestParser;
 import org.slf4j.Logger;
